requirements.txt
datetime 
pandas 
sqlalchemy 
requests 
py-cord
openai
mysql-connector-python
UserAgent
fake_useragent
logging